requirements.txt
 streamlit
  mysql-connector-python 
  pandas 
  requests 
  plotly